Profiles
So I finished up the quick and column filters and figured out the css but give me some input if you can, I want to put the dropdown in the same row as the quick filter but probably do that when everything else is working. I'm gonna be working on the finance page tomorrow to start my day and go from there! Signing off for the night!